--- src/Screens/ComingSoon/Movie/MovieInfo/MovieDescription/index.js
@@ -1,15 +1,0 @@
-import React from "react";
-import { Text } from "react-native";
-
-import { useOrientation } from "../../../../../Hooks/useOrientation";
-import { styles } from "./styles";
-
-export default function MovieDescription({ children }) {
-  const orientation = useOrientation();
-
-  return (
-    <Text style={orientation.isPortrait ? styles.portrait : styles.landscapet}>
-      {children}
-    </Text>
-  );
-}
