--- src/Screens/ComingSoon/Movie/MovieInfo/MovieDescription/styles.js
@@ -1,21 +1,0 @@
-import React from "react";
-import EStyleSheet from "react-native-extended-stylesheet";
-
-import { colors } from "../../../../../Styles/colors";
-
-export const styles = EStyleSheet.create({
-  portrait: {
-    marginTop: ".1rem",
-    marginHorizontal: ".5rem",
-    fontSize: ".9rem",
-    color: "grey",
-    lineHeight: "1.2rem",
-  },
-  landscapet: {
-    marginTop: ".1rem",
-    marginHorizontal: ".5rem",
-    fontSize: ".9rem",
-    color: "grey",
-    lineHeight: "1.2rem",
-  },
-});
