+++ src/components/Movie/SecundaryInfo/Action/styles.js
@@ -1,0 +1,22 @@
+import { StatusBar } from "react-native";
+import EStyleSheet from "react-native-extended-stylesheet";
+
+import { colors } from "../../../../Styles/colors";
+
+export const styles = EStyleSheet.create({
+  actionPortrait: {
+    alignItems: "center",
+  },
+  actionLandscapet: {
+    alignItems: "center",
+  },
+
+  textIconPortrait: {
+    fontSize: ".65rem",
+    color: "grey",
+  },
+  textIconLandscapet: {
+    fontSize: ".65rem",
+    color: "grey",
+  },
+});
