+++ src/components/Movie/SecundaryInfo/Rating/index.js
