+++ src/components/Movie/styles.js
